Extract author API from the substrate-rpc crate.
