+++ core/rpc/api/src/author/error.rs
@@ -16,8 +16,6 @@
 
 //! Authoring RPC module errors.
 
-use client;
-use transaction_pool::txpool;
 use crate::rpc;
 use crate::errors;
 
@@ -28,8 +26,10 @@ pub type Result<T> = std::result::Result<T, Error>;
 #[derive(Debug, derive_more::Display, derive_more::From)]
 pub enum Error {
 	/// Client error.
-	Client(client::error::Error),
+	#[display(fmt="Client error: {}", _0)]
+	Client(Box<dyn std::error::Error + Send>),
 	/// Transaction pool error,
+	#[display(fmt="Transaction pool error: {}", _0)]
 	Pool(txpool::error::Error),
 	/// Verification error
 	#[display(fmt="Extrinsic verification error: {}", _0)]
@@ -54,7 +54,7 @@ pub enum Error {
 impl std::error::Error for Error {
 	fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
 		match self {
-			Error::Client(ref err) => Some(err),
+			Error::Client(ref err) => Some(&**err),
 			Error::Pool(ref err) => Some(err),
 			Error::Verification(ref err) => Some(&**err),
 			_ => None,
