+++ core/rpc/api/src/author/hash.rs
