+++ core/rpc/api/src/errors.rs
