+++ core/rpc/api/src/metadata.rs
